[Refactor] #1 - Project Description을 위한 상수 정리, infoPlist 분리
